% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black76.R
\name{black76}
\alias{black76}
\title{black76}
\usage{
black76(Ft, K, type = c("C", "P"), Time, r, sigma)
}
\arguments{
\item{Ft}{Future price. Numeric object.}

\item{K}{Strike price. Numeric object.}

\item{type}{Type of the European option, call(C) or put(P). String object.}

\item{Time}{Time to maturity (in year). Numeric object.}

\item{r}{Continousely compounded interest rate. Numeric object.}

\item{sigma}{Annualsed volatility, square root of variance. Numeric object.}
}
\value{
List of 2 numeric objects, Option premium and Vega.
}
\description{
\code{black76} calculates the premium of European option on future based on black
model, Black(1976).
}
\examples{
black76(type = "C", Ft = 48.03, Time = 0.1423, K = 50, sigma = 0.2, r = 0.03)
black76(type = "P", Ft = 48.03, Time = 0.1423, K = 50, sigma = 0.2, r = 0.03)
}
