% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black76DF.R
\name{black76DF}
\alias{black76DF}
\title{black76DF}
\usage{
black76DF(df, Ftdf, sigmadf, Kdf, typedf, Timedf, rdf)
}
\arguments{
\item{df}{Name of a dataframe}

\item{Ftdf}{Name of a columns in the dataframe corresponding to future price.}

\item{sigmadf}{Name of a columns in the dataframe corresponding to
volatility.}

\item{Kdf}{Name of a columns in the dataframe corresponding to strike price.}

\item{typedf}{Name of a columns in the dataframe corresponding to option
type.}

\item{Timedf}{Name of a columns in the dataframe corresponding to interest
rate.}

\item{rdf}{Name of a columns in the dataframe corresponding to interest rate.}
}
\value{
Option premium calculated as a column in the dataframe.
}
\description{
\code{black76DF} appies \code{black76} to calculate the premium of European
option on future based on black model, Black(1976) on each row of a dataframe
with strike price, future price, option type, volatility, interest rate, and
time to maturity as columns.
}
\examples{
Ft <- 50
Kvec <- 80:120
Timevec <- c(c(3, 6, 9)/12, 1:5)
r <- 0.01
Vol <- runif(n = 328, min = 0.1, 0.3)
type <- sample(x = c("C", "P"), size = 328, replace = TRUE, prob = c(0.5, 0.5))
df <- data.frame(expand.grid(Kvec, Timevec), Ft, r, Vol, type)

df2 <- black76DF(df = df, Ftdf = 'Ft', sigmadf = 'Vol', Kdf = 'Var1',
                 typedf = 'type', rdf = 'r', Timedf = 'Var2')
                 head(df2)
}
