% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImpliedVol.R
\name{plotImpliedVol}
\alias{plotImpliedVol}
\title{plotImpliedVol}
\usage{
plotImpliedVol(df, Ftdf, Kdf, typedf, Timedf, rdf, optionPricedf,
  sigma_ini = NULL, tol = 1e-04, maxiter = 20, phi = 45, theta = 120)
}
\description{
\code{plotImpliedVol}
}
\examples{
example(black76ImVoDF)
df2 <- df2[sample(x = 1:nrow(df2), size = 10), ]
head(df2)
plotImpliedVol(df = df2, Ftdf = 'Ft', Kdf = 'Strike', typedf = 'type',
              Timedf = 'Time', rdf = 'r', optionPricedf = 'optionPrice',
              tol = 1e-5, maxiter = 1000, phi = 45, theta = 120)
}
